+++ addons/im_livechat/__manifest__.py
@@ -143,6 +143,7 @@ Help your customers with this chat, and analyse their feedback.
             'web/static/src/session.js',
             'web/static/src/legacy/js/core/session.js',
             'web/static/src/legacy/js/core/concurrency.js',
+            'web/static/src/legacy/js/core/cookie_utils.js',
             'web/static/src/legacy/js/core/utils.js',
             'web/static/src/legacy/js/core/dom.js',
             'web/static/src/legacy/js/core/qweb.js',
