+++ addons/web/__manifest__.py
@@ -58,6 +58,7 @@ This module provides the core of the Odoo Web Client.
             'web/static/src/legacy/js/promise_extension.js',
             'web/static/src/boot.js',
             'web/static/src/session.js',
+            'web/static/src/legacy/js/core/cookie_utils.js',
         ],
         'web.assets_common': [
             ('include', 'web._assets_helpers'),
@@ -75,6 +76,7 @@ This module provides the core of the Odoo Web Client.
             ('remove', 'web/static/src/legacy/js/promise_extension.js'),
             ('remove', 'web/static/src/boot.js'),
             ('remove', 'web/static/src/session.js'),
+            ('remove', 'web/static/src/legacy/js/core/cookie_utils.js'),
         ],
         'web.assets_backend': [
             ('include', 'web._assets_helpers'),
